Move pages out of /components and into /views
