--- assets/js/app.b9e2bd5e.js
@@ -1,2 +1,0 @@
-(function(t){function e(e){for(var r,a,s=e[0],u=e[1],c=e[2],p=0,f=[];p<s.length;p++)a=s[p],o[a]&&f.push(o[a][0]),o[a]=0;for(r in u)Object.prototype.hasOwnProperty.call(u,r)&&(t[r]=u[r]);l&&l(e);while(f.length)f.shift()();return i.push.apply(i,c||[]),n()}function n(){for(var t,e=0;e<i.length;e++){for(var n=i[e],r=!0,s=1;s<n.length;s++){var u=n[s];0!==o[u]&&(r=!1)}r&&(i.splice(e--,1),t=a(a.s=n[0]))}return t}var r={},o={1:0},i=[];function a(e){if(r[e])return r[e].exports;var n=r[e]={i:e,l:!1,exports:{}};return t[e].call(n.exports,n,n.exports,a),n.l=!0,n.exports}a.m=t,a.c=r,a.d=function(t,e,n){a.o(t,e)||Object.defineProperty(t,e,{enumerable:!0,get:n})},a.r=function(t){"undefined"!==typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(t,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(t,"__esModule",{value:!0})},a.t=function(t,e){if(1&e&&(t=a(t)),8&e)return t;if(4&e&&"object"===typeof t&&t&&t.__esModule)return t;var n=Object.create(null);if(a.r(n),Object.defineProperty(n,"default",{enumerable:!0,value:t}),2&e&&"string"!=typeof t)for(var r in t)a.d(n,r,function(e){return t[e]}.bind(null,r));return n},a.n=function(t){var e=t&&t.__esModule?function(){return t["default"]}:function(){return t};return a.d(e,"a",e),e},a.o=function(t,e){return Object.prototype.hasOwnProperty.call(t,e)},a.p="/assets/";var s=window["webpackJsonp"]=window["webpackJsonp"]||[],u=s.push.bind(s);s.push=e,s=s.slice();for(var c=0;c<s.length;c++)e(s[c]);var l=u;i.push([5,0]),n()})({"/9E1":function(t,e,n){},5:function(t,e,n){t.exports=n("Vtdi")},BDkB:function(t,e){},Curk:function(t,e,n){"use strict";var r=n("q8Yv"),o=n.n(r);o.a},R7CL:function(t,e,n){"use strict";var r=n("/9E1"),o=n.n(r);o.a},Vtdi:function(t,e,n){"use strict";n.r(e);var r={};n.r(r);var o={};n.r(o);n("VRzm");var i=n("Kw5r"),a=function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{attrs:{id:"app"}},[n("router-view",{key:t.$route.path})],1)},s=[],u=new i["a"],c=u,l="RECEIVE_PAGES",p="RECEIVE_SITE",f="UPDATE_APP",d={name:"app",components:{},data:function(){return{}},beforeCreate:function(){this.$store.commit(l,{pages:window.Kirby.pages}),this.$store.commit(p,{site:window.Kirby.site})},created:function(){window.addEventListener("resize",this.onResize,!0),window.addEventListener("scroll",this.onScroll),window.addEventListener("orientationchange",this.onOrientationChange),this.onResize()},destroyed:function(){window.removeEventListener("resize",this.onResize,!0),window.removeEventListener("scroll",this.onScroll),window.removeEventListener("orientationchange",this.onOrientationChange)},methods:{onResize:function(t){var e=window.innerWidth||document.documentElement.clientWidth||document.getElementsByTagName("body")[0].clientWidth;this.$store.commit(f,{field:"tablet",value:e<=960}),this.$store.commit(f,{field:"phone",value:e<=480}),c.$emit("app.resize",t)},onScroll:function(t){c.$emit("app.scroll",t)},onOrientationChange:function(t){c.$emit("app.orientationchange",t)}}},v=d,h=(n("nNx0"),n("KHd+")),m=Object(h["a"])(v,a,s,!1,null,null,null),g=m.exports,b=n("jE9Z"),y=function(){var t=this,e=t.$createElement;t._self._c;return t._m(0)},w=[function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"home"},[n("div",{staticClass:"content-container"},[t._v("\n    Hi\n  ")])])}],_={name:"home",components:{},data:function(){return{}}},E=_,j=(n("vKOL"),Object(h["a"])(E,y,w,!1,null,null,null)),O=j.exports,C=function(){var t=this,e=t.$createElement;t._self._c;return t._m(0)},$=[function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"about"},[n("div",{staticClass:"content-container"})])}],P={name:"about",components:{},data:function(){return{}}},x=P,L=(n("Curk"),Object(h["a"])(x,C,$,!1,null,null,null)),B=L.exports,R=function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"project"},[n("div",{staticClass:"content-container",domProps:{innerHTML:t._s(t.page.text)}})])},S=[],T={name:"project",props:["uid"],components:{},computed:{page:function(){return this.$store.getters.getPageByUID(this.uid)}},data:function(){return{}}},k=T,z=(n("sduy"),Object(h["a"])(k,R,S,!1,null,null,null)),U=z.exports,D=function(){var t=this,e=t.$createElement,n=t._self._c||e;return n("div",{staticClass:"projects"},[n("div",{staticClass:"content-container"},[n("ul",t._l(t.subpages,function(e){return n("li",[n("router-link",{attrs:{to:{name:e.type,params:{uid:e.uid}}}},[t._v(t._s(e.title))])],1)}))])])},I=[],K={name:"projects",components:{},computed:{subpages:function(){return this.$store.getters.getPagesByType("project")}},data:function(){return{}}},W=K,M=(n("R7CL"),Object(h["a"])(W,D,I,!1,null,null,null)),V=M.exports;i["a"].use(b["a"]);var J=new b["a"]({mode:"history",routes:[{path:"/",name:"home",component:O},{path:"/projects",name:"projects",component:V},{path:"/project/:uid",name:"project",component:U,props:!0},{path:"/about",name:"about",component:B},{path:"*",redirect:"/"}]}),A=n("L2JU"),F=n("BDkB"),H=(n("sFTC"),[]),N=n("oyJW"),q={pages:[]},Y={getFileByURL:function(t,e){return function(n){var r=arguments.length>1&&void 0!==arguments[1]?arguments[1]:t.pages,o=r.length;while(o--){var i=r[o].files.length;while(i--)if(r[o].files[i].url===n)return r[o].files[i];var a=e.getFileByURL(n,r[o].children);if(a)return a}}},getPagesByType:function(t,e){return function(n){var r=arguments.length>1&&void 0!==arguments[1]?arguments[1]:t.pages,o=[];if(!r||!r.length)return o;for(var i=0;i<r.length;i++){r[i].type===n&&o.push(r[i]);var a=e.getPagesByType(n,r[i].children);o=o.concat(a)}return o}},getPageByUID:function(t,e){return function(n){var r=arguments.length>1&&void 0!==arguments[1]?arguments[1]:t.pages,o=r.length;while(o--){if(r[o].uid===n)return r[o];var i=e.getPageByUID(n,r[o].children);if(i)return i}}}},G={},Z=Object(N["a"])({},l,function(t,e){var n=e.pages;t.pages=n}),Q={state:q,getters:Y,actions:G,mutations:Z},X={site:{}},tt={},et={},nt=Object(N["a"])({},p,function(t,e){var n=e.site;t.site=n}),rt={state:X,getters:tt,actions:et,mutations:nt},ot={tablet:!1,phone:!1},it={},at={},st=Object(N["a"])({},f,function(t,e){var n=e.field,r=e.value;i["a"].set(t,n,r)}),ut={state:ot,getters:it,actions:at,mutations:st},ct=!1;i["a"].use(A["a"]);var lt=new A["a"].Store({mutations:F,actions:r,getters:o,modules:{app:ut,pages:Q,site:rt},strict:ct,plugins:H});i["a"].config.productionTip=!1,new i["a"]({router:J,store:lt,render:function(t){return t(g)}}).$mount("#app")},"W/Ot":function(t,e,n){},boi5:function(t,e,n){},mtsb:function(t,e,n){},nNx0:function(t,e,n){"use strict";var r=n("boi5"),o=n.n(r);o.a},q8Yv:function(t,e,n){},sduy:function(t,e,n){"use strict";var r=n("mtsb"),o=n.n(r);o.a},vKOL:function(t,e,n){"use strict";var r=n("W/Ot"),o=n.n(r);o.a}});
-//# sourceMappingURL=app.b9e2bd5e.js.map
