+++ src/router/index.js
@@ -1,9 +1,9 @@
 import Vue from 'vue'
 import Router from 'vue-router'
-import Home from '@/components/Home'
-import About from '@/components/About'
-import Project from '@/components/Project'
-import Projects from '@/components/Projects'
+import Home from '@/views/Home'
+import About from '@/views/About'
+import Project from '@/views/Project'
+import Projects from '@/views/Projects'
 
 Vue.use(Router)
